backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
pymupdf==1.23.8
python-docx==1.1.0
requests==2.31.0
sqlite3
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
python-dotenv==1.0.0
aiofiles==23.2.1
pypandoc==1.11
pylatex==1.4.2
smtplib
email
schedule==1.2.0
